buffer overrun bug

+++ Common/DtaDev.h
@@ -113,7 +113,7 @@ public:
 	 * @param bufferlen length of the input/output buffer
 	 */
 	virtual uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-		void * buffer, uint32_t bufferlen) = 0;
+		void * buffer, size_t bufferlen) = 0;
 	/** OS specific command to Wait for specified number of milliseconds
 	 * @param milliseconds  number of milliseconds to wait
 	 */
