+++ Common/DtaDiskType.cpp
@@ -29,7 +29,7 @@ DtaDiskType::~DtaDiskType() {}
 #endif
 void DtaDiskType::init(const char * devref) {}
 uint8_t	DtaDiskType::sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-	void * buffer, uint32_t bufferlen) {
+	void * buffer, size_t bufferlen) {
 	return 10;
 }
 void DtaDiskType::identify(DTA_DEVICE_INFO& disk_info) {}
