+++ macos/DtaDevMacOSTPer_SAT.h
@@ -50,7 +50,7 @@ public:
      * @param bufferlen length of the input/output buffer
      */
     uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-                    void * buffer, uint16_t bufferlen);
+                    void * buffer, size_t bufferlen);
     
     /** Connect to "user client" (driver shim in userland)
      */
