+++ macos/DtaDevOS.h
@@ -65,7 +65,7 @@ public:
          * @param bufferlen length of the input/output buffer
          */
     uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-                    void * buffer, uint32_t bufferlen);
+                    void * buffer, size_t bufferlen);
     /** A static class to scan for supported drives */
     static int diskScan();
